move saifp model to sentry_apps
