+++ src/sentry/models/authprovider.py
@@ -122,11 +122,11 @@ class AuthProvider(ReplicatedControlModel):
         return get_scim_token(self.flags.scim_enabled, self.organization_id, self.provider)
 
     def enable_scim(self, user):
-        from sentry.models.integrations.sentry_app_installation_for_provider import (
-            SentryAppInstallationForProvider,
-        )
         from sentry.sentry_apps.logic import SentryAppCreator
         from sentry.sentry_apps.models.sentry_app_installation import SentryAppInstallation
+        from sentry.sentry_apps.models.sentry_app_installation_for_provider import (
+            SentryAppInstallationForProvider,
+        )
 
         if (
             not self.get_provider().can_use_scim(self.organization_id, user)
@@ -186,7 +186,7 @@ class AuthProvider(ReplicatedControlModel):
 
     def disable_scim(self):
         from sentry import deletions
-        from sentry.models.integrations.sentry_app_installation_for_provider import (
+        from sentry.sentry_apps.models.sentry_app_installation_for_provider import (
             SentryAppInstallationForProvider,
         )
